Make Emoji Search's Compose a bit more idiomatic
Much of its structure was retained from before we had Compose when it was in the Zipline repo. But in Compose, all of the logic can happen directly on state.
