--- samples/emoji-search/presenters/src/commonMain/kotlin/app/cash/zipline/samples/emojisearch/models.kt
@@ -1,39 +1,0 @@
-/*
- * Copyright (C) 2022 Square, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package app.cash.zipline.samples.emojisearch
-
-import example.values.TextFieldState
-import kotlinx.serialization.Serializable
-
-@Serializable
-sealed class EmojiSearchEvent {
-  @Serializable
-  class SearchTermEvent(
-    val searchTerm: TextFieldState,
-  ) : EmojiSearchEvent()
-}
-
-@Serializable
-data class EmojiSearchViewModel(
-  val searchTerm: TextFieldState,
-  val images: List<EmojiImage>,
-)
-
-@Serializable
-data class EmojiImage(
-  val label: String,
-  val url: String,
-)
